Simulation: use btree for storing relationships, use iterators over loops
I did a lot of things here:
most importantly, I added a btree with a custom enum Key type that sorts keys in
useful orders:
- all static bodies (top level bodies like the sun) are in a contiguous range,
making it easy to use the tree to retrieve them without having to iterate over
all bodies.
- all direct satellites of a body are in a contiguous range, making it easy to
use the tree to retrieve them (it was reasonably easy before, too).
- children have a special parent node linking to the parent.

not having to touch the SimulatedBody struct to look up parents and children
should be more performant (insignificantly so) and more ergonomic (in my opinion)

For perfect rusty, idiomatic code, there should probably be another enum for the
Value of the tree, differentiating between ids and indices, but that would make
the type twice as big!